+++ src/lib.rs
@@ -14,8 +14,11 @@ pub mod crypto;
 /// Steganography module containing different implementations of encoding methods.
 pub mod steganography;
 
+/// Image API contains an HTTP client to fetch images from a remote API
+pub mod image_api;
+
 use thiserror::Error;
-#[derive(Error, Debug, PartialEq)]
+#[derive(Error, Debug, Eq, PartialEq)]
 pub enum StegError {
     #[error("Encoded message not found in data")]
     EncodingNotFound,
@@ -31,7 +34,7 @@ pub enum StegError {
     Unknown,
 }
 
-#[derive(Error, Debug, PartialEq)]
+#[derive(Error, Debug, Eq, PartialEq)]
 pub enum CryptoError {
     #[error("Failed to get random salt")]
     Salt,